requirements.txt
langgraph
langchain
python-dotenv
langchain-google-genai
flask
flask-cors